Idempotent build scripts
This fixes #2 and generates the index.html file used on the server in the same
way it is generated for production builds. This avoids the issue of the dev
index.html conflicting with production builds and avoids issues with version
control of index.html in `build/` and accidental commits of changes to it.