Refactored options to have meta-data
Choice options are now represented as an object instead of just an array of
consequences. The consequences are now stored on a key called 'consequences'.
This is a change to prep for the addition of restrictions on the choices which
require more information be stored on the option than just the outcome.